토픽 생성/수정/삭제 전 팝업창

+++ src/main/resources/static/javascript/TopicSetting.js
@@ -24,60 +24,38 @@ $(document).ready(function () {
 
     $("#new").click(function () {
         $("#popup_layer_new").css("display", "block");
-        $("form").submit(function (e) {
-            e.preventDefault(); // avoid to execute the actual submit of the form.
-
-            if ($("#inputTopic").val().trim() == "" || $("#inputMonitoring").val().trim() == ""
-                || $("#inputIP").val().trim() == "" || $("#inputPort").val().trim() == "") {
-                alert("값을 입력해주세요.");
-            } else {
-                let form = $(this);
-                $.ajax({
-                    type: "POST",
-                    url: "saveTopic",
-                    data: form.serialize(), // serializes the form's elements.
-                    success: function (result) {
-                        console.log(result);
-                        $("#popup_layer_new").css("display", "none");
-                        $("#page").load("TopicSetting");
-                    },
-                    error: function (error) {
-                        console.log("실패:" + error.status);
-                        if (error.status == 500) {
-                            alert("이미 존재하는 토픽명입니다.");
-                        }
-                    }
-                });
-            }
-        });
 
     });
+
     $("#delete").click(function () {
         if (num == "") {
             alert("항목을 선택해주세요.");
         } else {
-            $.ajax({
-                url: "getTopic/" + num,
-                type: "GET",
-                success: function (result) {
-                    console.log(result);
-                    if (result.status == "Running") {
-                        alert("구독 중에는 삭제할 수 없습니다.");
-                    } else {
-                        $.ajax({
-                            url: "delete/" + num,
-                            type: "DELETE",
-                            success: function (result) {
-                                console.log(result);
-                                alert("삭제되었습니다.");
-                                $("#page").load("TopicSetting");
-                            }
-                        })
+            if (confirm("삭제하시겠습니까?")) {
+                $.ajax({
+                    url: "getTopic/" + num,
+                    type: "GET",
+                    success: function (result) {
+                        console.log(result);
+                        if (result.status == "Running") {
+                            alert("구독 중에는 삭제할 수 없습니다.");
+                        } else {
+                            $.ajax({
+                                url: "delete/" + num,
+                                type: "DELETE",
+                                success: function (result) {
+                                    console.log(result);
+                                    alert("삭제되었습니다.");
+                                    $("#page").load("TopicSetting");
+                                }
+                            })
+                        }
                     }
-                }
-            })
+                })
+            }
         }
     });
+
     $("#edit").click(function () {
         if (num == "") {
             alert("항목을 선택해주세요.");
@@ -91,53 +69,82 @@ $(document).ready(function () {
                         alert("구독 중에는 편집할 수 없습니다.");
                     } else {
                         $("#popup_layer_edit").css("display", "block");
-                        $.ajax({
-                            url: "getTopic/" + num,
-                            type: "GET",
-                            success: function (result) {
-                                console.log(result);
-                                $("#editTopic").val(result.topicName);
-                                $("#editMonitoring").val(result.monitoringName);
-                                $("#editIP").val(result.ip);
-                                $("#editPort").val(result.port);
-
-                            }
-                        })
-                        $("form").submit(function (e) {
-                            e.preventDefault(); // avoid to execute the actual submit of the form.
-
-                            if ($("#editTopic").val().trim() == "" || $("#editMonitoring").val().trim() == ""
-                                || $("#editIP").val().trim() == "" || $("#editPort").val().trim() == "") {
-                                alert("값을 입력해주세요.");
-                            } else {
-                                let form = $(this);
-
-                                $.ajax({
-                                    type: "POST",
-                                    url: "edit/" + num,
-                                    data: form.serialize(), // serializes the form's elements.
-                                    success: function (result) {
-                                        console.log(result);
-                                        $("#popup_layer_edit").css("display", "none");
-                                        $("#page").load("TopicSetting");
-                                    },
-                                    error: function (error) {
-                                        console.log("실패:" + error.status);
-                                        if (error.status == 500) {
-                                            alert("이미 존재하는 토픽명입니다.");
-                                        }
-                                    }
-                                });
-                            }
-                        });
+                        $("#editTopic").val(result.topicName);
+                        $("#editMonitoring").val(result.monitoringName);
+                        $("#editIP").val(result.ip);
+                        $("#editPort").val(result.port);
                     }
                 }
             })
         }
     });
+
+    $("#saveBtn").click(function () {
+        if (confirm("저장하시겠습니까?")) {
+            if ($("#inputTopic").val().trim() == "" || $("#inputMonitoring").val().trim() == ""
+                || $("#inputIP").val().trim() == "" || $("#inputPort").val().trim() == "") {
+                alert("값을 입력해주세요.");
+            } else {
+                $.ajax({
+                    type: "POST",
+                    url: "saveTopic",
+                    data: {
+                        topicName: $("#inputTopic").val(),
+                        monitoringName: $("#inputMonitoring").val(),
+                        ip: $("#inputIP").val(),
+                        port: $("#inputPort").val()
+                    },
+                    success: function (result) {
+                        console.log(result);
+                        $("#popup_layer_new").css("display", "none");
+                        $("#page").load("TopicSetting");
+                    },
+                    error: function (error) {
+                        console.log("실패:" + error.status);
+                        if (error.status == 500) {
+                            alert("이미 존재하는 토픽명입니다.");
+                        }
+                    }
+                })
+            }
+        }
+    });
+
+    $("#editBtn").click(function () {
+        if (confirm("수정하시겠습니까?")) {
+            if ($("#editTopic").val().trim() == "" || $("#editMonitoring").val().trim() == ""
+                || $("#editIP").val().trim() == "" || $("#editPort").val().trim() == "") {
+                alert("값을 입력해주세요.");
+            } else {
+                $.ajax({
+                    type: "POST",
+                    url: "edit/"+num,
+                    data: {
+                        topicName: $("#editTopic").val(),
+                        monitoringName: $("#editMonitoring").val(),
+                        ip: $("#editIP").val(),
+                        port: $("#editPort").val()
+                    },
+                    success: function (result) {
+                        console.log(result);
+                        $("#popup_layer_edit").css("display", "none");
+                        $("#page").load("TopicSetting");
+                    },
+                    error: function (error) {
+                        console.log("실패:" + error.status);
+                        if (error.status == 500) {
+                            alert("이미 존재하는 토픽명입니다.");
+                        }
+                    }
+                })
+            }
+        }
+    });
+
     $("#saveCancelBtn").click(function () {
         $("#popup_layer_new").css("display", "none");
     });
+
     $("#editCancelBtn").click(function () {
         $("#popup_layer_edit").css("display", "none");
     });
